Bugfix for post destructor update calls
Update was called after GameObject.Destroy() but before it was finished
with the destructor which could cause a null reference as happened in
MainGunProjectile.

Fixed by doing an alive check in the update loop and locking down the
GameObject.Alive property to only be settable from the GameObject - it
now manages its own lifecycle better.