The Mega Update
Moved UserContentServer to correct folder
Added Linux as a platform.
Upgraded to VS2015 RC (and started using some of the c# 6 features such
as string interpolation)
Moved around settings
Refactored mod engine to allow for *.mod files + avoided extraneous
loading
Added debug flags so that we don't load in a separate appdomain in debug
mode (which makes debugging a PITA) but still do so in release.
Wrapped try{}'s in Mod Loader in #if !DEBUG's so that errors are easy to
find in debug mode.