NON WORKING COMMIT
plus, added a better house texture.
Fixed Setting to string, so it now is prints the value of the settings.
Added an initialization note to menus so errors are always recorded.
Switched global exception catch to only hook when not in debug mode

Also, visual studio is not being nice. The new largehouse texture breaks
everything. For some reason, the app silent crashes and the exception is
not catchable...Visual studio doesn't see it and try-catch doesn't
either.