Many things
Fix fullgamestate implementation to not send entire map over the
network.
Set mod database to work on major version keys too, not just names.
Added follow camera to the tanks. The faster you go, the more the camera
zooms out and the more ahead of you it goes.

Also, clean up networkedgame to work better.

Added ModAssetInfo and ModInfo structs, which describe a mod so that it
can be located in the database or (with modassetinfo) describes a single
asset that can be located.