More "a bunch of stuff"'s
Added a pause menu
Made all the UI pages share Karmatic Arcade font
Standardized a bit of the ui design for prettiness
Removed the OldPage class
Fixed input driver locking mouse
Made mouse visible on menus
Changed a few of the error messages
Fixed a bug with users leaving GameCore before the game starts (which
means team is null)

Set Connection timeout to 15, rather than infinite for debug mode

Fixed a bug in the network layer:
If the user was not the host, they would not send any ActionBase objects
during login. Thus, the static constructor which added all of the
reflection information was never called. So, when the server sent a
message, it was unable to deserialize, as there was no information
available. Fixed by moving reflection into Network processor, which
receives the messages.

Added more error handling to NetworkProcessor